requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
blinker==1.9.0
boto3==1.35.88
botocore==1.35.99
cffi==1.17.1
click==8.2.1
cryptography==41.0.7
dotenv==0.9.9
flasgger==0.9.7.1
Flask==3.1.1
flask-cors==6.0.1
frozenlist==1.7.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
multidict==6.5.0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyJWT==2.8.0
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.26.0
s3transfer==0.10.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
